Add strategy support for state-archives
